requirements.txt
# Requirements for Emotion-Aware Chatbot — Aria
# Install: pip install -r requirements.txt

# ── Web UI ────────────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.18.0

# ── LLM ──────────────────────────────────────────────────────────────────────
anthropic>=0.25.0

# ── NLP / Sentiment ───────────────────────────────────────────────────────────
transformers>=4.38.0
torch>=2.0.0
vaderSentiment>=3.3.2
tokenizers>=0.15.0
sentencepiece>=0.1.99
accelerate>=0.26.0

# ── AWS ───────────────────────────────────────────────────────────────────────
boto3>=1.34.0

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

python-dotenv>=1.0.0        # Load .env file locally